% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rc_array.R
\name{make_rc_array}
\alias{make_rc_array}
\title{Make reverse correlation array}
\usage{
make_rc_array(rc_list, image = "combined", smoothingK = 0)
}
\arguments{
\item{rc_list}{a reverse correlation list imported from .Rda format}

\item{image}{the image file to be used. Has to be one in c('combined', 'ci', 'base', 'scaled'). Default is 'combined'.}

\item{smoothinkK}{the full-width at half max Gaussian smoothing kernel to be applied to data. Default is no smoothing.}
}
\description{
Converts list to 3D array object
}
\examples{
make_rc_array(rda_import_list, smoothingK = 10)
}
\author{
Robert S. Chavez
}
