% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masker.R
\name{masker}
\alias{masker}
\title{Matrix masking function}
\usage{
masker(matrix, as_vector = FALSE)
}
\arguments{
\item{matrix}{binary matrix of the mask to be used}

\item{as_vector}{return a vector instead of a matrix}
}
\description{
Converts 0 values of a binary matrix into NAs for analysis and plotting
}
\examples{
mat <- matrix(sample(0:1, 100, replace = TRUE),nrow = 10, ncol = 10 )
masker(mat)
}
\author{
Robert S. Chavez
}
