% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_array_plot.R
\name{rc_array_plot}
\alias{rc_array_plot}
\title{Plot 2D image from array}
\usage{
rc_array_plot(stat_array, palette = "Greys", legend = FALSE)
}
\arguments{
\item{stat_array}{a 2D matrix (e.g., from stat_array())}

\item{palette}{character for color palette from RColorBrewer}

\item{legend}{logical to include legend}
}
\description{
Returns a ggplot2 plot object showing an image of a 2D input. Can be useful for quality control.
}
\examples{

rc_array_plot(my_mean_rc_image)
}
\author{
Robert S. Chavez
}
