% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_corrp_plot.R
\name{rc_corrp_plot}
\alias{rc_corrp_plot}
\title{Plot corrected p-value reverse correlation maps}
\usage{
rc_corrp_plot(stat_df, background_image = NULL, corrp_thresh = 0.05)
}
\arguments{
\item{stat_df}{a data frame object with t-value, p-value, and coordinate information (e.g., output from run_pixel_ttests())}

\item{background_image}{a matrix in the same dimensions as the stat_df}

\item{corrp_thresh}{corrected p-value threshold}
}
\description{
Returns a ggplot2 plot object corrected p-values spatial statistical map
}
\examples{
rc_corrp_plot(my_pixel_ttest, background_image = my_mean_rc_image)
}
\author{
Robert S. Chavez
}
