% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_corstat_plot.R
\name{rc_corstat_plot}
\alias{rc_corstat_plot}
\title{Plot correlation coefficients within significant corrected p-value reverse correlation maps}
\usage{
rc_corstat_plot(stat_df, background_image = NULL, corrp_thresh = 0.05)
}
\arguments{
\item{stat_df}{a data frame object with correlation coefficients, p-value, and coordinate information (e.g., output from run_pixel_correlation())}

\item{background_image}{a matrix in the same dimensions as the stat_df}

\item{corrp_thresh}{corrected p-value threshold}
}
\description{
Returns a ggplot2 plot object with correlation coefficient within corrected p-value areas spatial statistical map
}
\examples{
rc_corstat_plot(my_pixel_cortest, background_image = my_mean_rc_image)
}
\author{
Robert S. Chavez
}
