% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_tstat_plot.R
\name{rc_tstat_plot}
\alias{rc_tstat_plot}
\title{Plot t-values within significant corrected p-value reverse correlation maps}
\usage{
rc_tstat_plot(stat_df, backgroud_image = NULL, corrp_thresh = 0.05)
}
\arguments{
\item{stat_df}{a data frame object with t-value, p-value, and coordinate information (e.g., output from run_pixel_ttests())}

\item{backgroud_image}{a matrix in the same dimensions as the stat_df}

\item{corrp_thresh}{corrected p-value threshold}
}
\description{
Returns a ggplot2 plot object showing t-values of significant spatial statistical maps after correcting for multiple comparisons.
}
\examples{

rc_tstat_plot(my_pixel_ttest, background_image = my_mean_rc_image)
}
\author{
Robert S. Chavez
}
