% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_txt.R
\name{read_image_txt}
\alias{read_image_txt}
\title{Read text image matrix file}
\usage{
read_image_txt(path)
}
\arguments{
\item{path}{path string to file}
}
\description{
Reads a 2D text file and converts it into a matrix. Useful for masking.
}
\examples{

read_image_txt('~/my_mask_file.txt')
}
\author{
Robert S. Chavez
}
