% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rc_rda.R
\name{read_rc_rda}
\alias{read_rc_rda}
\title{Read reverse correlation .Rda file}
\usage{
read_rc_rda(path)
}
\arguments{
\item{path}{path string to file}
}
\description{
Reads .Rda file and converts it into a list for analysis.
}
\examples{

read_rc_rda('~/my_rc_data.Rda')
}
\author{
Robert S. Chavez
}
