% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pixel_correlation.R
\name{run_pixel_correlation}
\alias{run_pixel_correlation}
\title{Runs correlation between a variable and a reverse correlation 3D array}
\usage{
run_pixel_correlation(
  array,
  cor_var,
  mask = NULL,
  cor_method = "pearson",
  corr_p_method = "fdr"
)
}
\arguments{
\item{array}{a reverse correlation 3D array}

\item{cor_var}{a vector of values to correlate with pixels}

\item{mask}{2D matrix used for selecting pixels to analyze (i.e., "masking")}

\item{cor_method}{correlation method to use, either 'pearson', 'spearman', or 'kendall'}

\item{corr_p_method}{multiple comparison correction method from p.adjust() to be used}
}
\description{
Calculates 2D summary statistics from all points of a reverse correlation list object
}
\examples{
run_pixel_correlation(facearray, bias_variable, mask = NULL, cor_method = 'pearson', corr_p_method = 'fdr' ) 
}
\author{
Robert S. Chavez
}
