% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pixel_ttests.R
\name{run_pixel_ttests}
\alias{run_pixel_ttests}
\title{Runs t-test between two reverse correlation 3D arrays}
\usage{
run_pixel_ttests(
  array1,
  array2,
  mask = NULL,
  corr_p_method = "fdr",
  paired = FALSE,
  var_equal = FALSE
)
}
\arguments{
\item{array1}{a reverse correlation 3D array}

\item{array2}{a reverse correlation 3D array}

\item{mask}{2D matrix used for selecting pixels to analyze (i.e., "masking")}

\item{corr_p_method}{multiple comparison correction method from p.adjust() to be used}

\item{paired}{logical if the t-test is a paired samples t-test}

\item{var_equal}{logical if equal variance is assumed}
}
\description{
Calculates 2D summary statistics from all points of a reverse correlation list object
}
\examples{
run_pixel_ttests(array1, array2, mask = maskface, paired = T)
}
\author{
Robert S. Chavez
}
