% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_array.R
\name{stat_array}
\alias{stat_array}
\title{Calculate summary statistics from reverse correlation images}
\usage{
stat_array(rc_list, stat_func = "mean", image = "combined", smoothingK = 0)
}
\arguments{
\item{rc_list}{a reverse correlation list imported from .Rda format}

\item{stat_func}{a character of the name of the function to be used. Default is 'mean'.}

\item{image}{the image file to be used. Has to be one in c('combined', 'ci', 'base', 'scaled'). Default is 'combined'.}

\item{smoothinkK}{the full-width at half max Gaussian smoothing kernel to be applied to data. Default is no smoothing.}
}
\description{
Calculates 2D summary statistics from all points of a reverse correlation list object
}
\examples{
stat_array(rda_import) # will output the average image from 3D array
}
\author{
Robert S. Chavez
}
